+++ meshtaichi_patcher/mesh.py
@@ -1,3 +1,4 @@
+from meshtaichi_patcher_core import Patcher_cpp
 from .relation import Relation
 import numpy as np
 
@@ -66,36 +67,50 @@ class Mesh:
         return self.relation[(from_end, to_end)]
     
     def patch(self, cluster):
-        self.cluster = cluster
-        self.owned = []
+        self.patcher = Patcher_cpp()
         for order in range(self.n_order - 1):
-            color = np.zeros([self.get_size(order)], dtype=np.int32)
-            relation = self.get_relation(self.n_order - 1, order)
-            for u in range(self.get_size(self.n_order - 1)):
-                for v in relation[u]:
-                    color[v] = cluster.color[u]
-            owned = [[] for i in cluster.patch]
-            for v, c in enumerate(color):
-                owned[c].append(v)
-            self.owned.append(Relation(owned))
-        owned = None
-        self.owned.append(Relation(cluster.patch))
+            self.get_relation(self.n_order - 1, order)
+        for order in range(self.n_order):
+            self.get_relation(0, order)
+        for i, j in self.relation:
+            self.patcher.set_relation(i, j, self.relation[(i, j)].csr)
+        self.patcher.n_order = self.n_order
+        self.patcher.patch(Relation(cluster.patch).csr)
+        self.owned = []
         self.total = []
         for order in range(self.n_order):
-            total = []
-            for c, p in enumerate(cluster.patch):
-                s = set(self.owned[order][c])
-                rel_0 = self.get_relation(self.n_order - 1, 0)
-                rel_1 = self.get_relation(0, order)
-                l = []
-                for u in p:
-                    for v in rel_0[u]:
-                        for w in rel_1[v]:
-                            if w not in s:
-                                s.add(w)
-                                l.append(w)
-                total.append(list(self.owned[order][c]) + l)
-            self.total.append(Relation(total))
+            self.owned.append(Relation(self.patcher.get_owned(order)))
+            self.total.append(Relation(self.patcher.get_total(order)))
+        # self.cluster = cluster
+        # self.owned = []
+        # for order in range(self.n_order - 1):
+        #     color = np.zeros([self.get_size(order)], dtype=np.int32)
+        #     relation = self.get_relation(self.n_order - 1, order)
+        #     for u in range(self.get_size(self.n_order - 1)):
+        #         for v in relation[u]:
+        #             color[v] = cluster.color[u]
+        #     owned = [[] for i in cluster.patch]
+        #     for v, c in enumerate(color):
+        #         owned[c].append(v)
+        #     self.owned.append(Relation(owned))
+        # owned = None
+        # self.owned.append(Relation(cluster.patch))
+        # self.total = []
+        # for order in range(self.n_order):
+        #     total = []
+        #     for c, p in enumerate(cluster.patch):
+        #         s = set(self.owned[order][c])
+        #         rel_0 = self.get_relation(self.n_order - 1, 0)
+        #         rel_1 = self.get_relation(0, order)
+        #         l = []
+        #         for u in p:
+        #             for v in rel_0[u]:
+        #                 for w in rel_1[v]:
+        #                     if w not in s:
+        #                         s.add(w)
+        #                         l.append(w)
+        #         total.append(list(self.owned[order][c]) + l)
+        #     self.total.append(Relation(total))
     
     def get_element_meta(self, order):
         ans = {}
@@ -118,34 +133,43 @@ class Mesh:
         ans = {}
         ans["from_order"] = from_end
         ans["to_order"] = to_end
-        total_from = self.total[from_end] if from_end > to_end else self.owned[from_end]
-        total_to = self.total[to_end]
-        relation = self.get_relation(from_end, to_end)
-        offset = []
-        value = []
-        patch_off = [0]
-        for i in range(len(total_from)):
-            d = {}
-            for j, u in enumerate(total_to[i]):
-                d[u] = j
-            matrix = []
-            for u in total_from[i]:
-                l = []
-                for v in relation[u]:
-                    l.append(d[v] if v in d else 0)
-                matrix.append(l)
-            rel = Relation(matrix)
-            offset = offset + list(rel.offset)
-            value = value + list(rel.value)
-            patch_off.append(patch_off[-1] + rel.offset[-1])
-        ans["offset"], ans["value"] = np.array(offset), np.array(value)
-        ans["patch_offset"] = np.array(patch_off[:-1])
+        self.patcher.set_relation(from_end, to_end, self.get_relation(from_end, to_end).csr)
+        csr = self.patcher.get_relation_meta(from_end, to_end)
+        ans["offset"], ans["value"] = csr.offset, csr.value
+        ans["patch_offset"] = self.patcher.get_patch_offset(from_end, to_end)
         return ans
+        # ans = {}
+        # ans["from_order"] = from_end
+        # ans["to_order"] = to_end
+        # total_from = self.total[from_end] if from_end > to_end else self.owned[from_end]
+        # total_to = self.total[to_end]
+        # relation = self.get_relation(from_end, to_end)
+        # offset = []
+        # value = []
+        # patch_off = [0]
+        # for i in range(len(total_from)):
+        #     d = {}
+        #     for j, u in enumerate(total_to[i]):
+        #         d[u] = j
+        #     matrix = []
+        #     for u in total_from[i]:
+        #         l = []
+        #         for v in relation[u]:
+        #             l.append(d[v] if v in d else 0)
+        #         matrix.append(l)
+        #     rel = Relation(matrix)
+        #     offset = offset + list(rel.offset)
+        #     value = value + list(rel.value)
+        #     patch_off.append(patch_off[-1] + rel.offset[-1])
+        # ans["offset"], ans["value"] = np.array(offset), np.array(value)
+        # ans["patch_offset"] = np.array(patch_off[:-1])
+        # return ans
     
     def get_meta(self, relations):
         c2i = {'v': 0, 'e': 1, 'f': 2, 'c': 3}
         ans = {}
-        ans["num_patches"] = len(self.cluster.patch)
+        # ans["num_patches"] = len(self.cluster.patch)
+        ans["num_patches"] = len(self.owned[-1])
         ans["elements"] = []
         for order in range(self.n_order):
             ans["elements"].append(self.get_element_meta(order))
