+++ src/patcher_api.h
@@ -1,6 +1,6 @@
 #include <vector>
 #include <string>
-#include "patcher.h"
+#include "patcher_old.h"
 
 namespace MeshTaichi {
 std::string run_mesh(std::string mesh_name, std::vector<std::string> relations);
